Gas tracker (#350)
* fix (docker compose): fix Relayer address in bridge docker compose

* feature (trustless-evm-bridge): fee storage for relayer

* feature (trustless-evm-bridge): BatchDispatched eth event to identify relayer

* fix (bridge-docker): fix bugs in evm bridge docker:
- obsolete account Bob changed to Relayer
- added missing `jq` dependency to Sora Dockerfile

* fix (evm-trusted-bridge): fix nonce in sub-relayer

* feature (evm-trusted-bridge): add gas tracking to contract

* feature (evm-trusted-bridge): eth storage for gas proof

+++ pallets/trustless-bridge/bridge-proxy/src/mock.rs
@@ -218,7 +218,7 @@ impl dispatch::Config for Test {
 
 parameter_types! {
     pub const MaxMessagePayloadSize: u64 = 2048;
-    pub const MaxMessagesPerCommit: u64 = 3;
+    pub const MaxMessagesPerCommit: u8 = 3;
     pub const Decimals: u32 = 12;
 }
 pub struct FeeConverter;
